Docker image to compile the Go project artifacts
This patch provides the Docker image that include required tools
to compile the Go binaries, manage the Go project dependecies as
well as generate the swagger documentation and GRPC servers based
on protocol buffers schema.

To build the image simply use the ```make``` command, that will
update the images of all versions.